Correction of a mistake in losses during pipeline transport of natural gas
In the model for losses of ch4 during pipeline transport of natural gas,
there was an error in modelling, as a consequence there were wrong
derived relations (not only CH4 direct emissions, but also other GHGs).
Also, added combustion and physical properties of anthracite, and of
some other fuels. Processes of combustion per LHV / HHV energy units are
now available for the same fuels.